requirements.txt
brotlipy==0.2.0
cffi==1.5.2
Django==1.9.5
djangorestframework==3.3.3
hashids==1.1.0
Markdown==2.6.6
pycparser==2.14
whitenoise==3.0
